web/requirements.txt
Flask[async]==3.0.2
opensearch-py==2.8.0
python-decouple==3.8
requests==2.32.3
pandas==2.2.3
pyarrow==19.0.1
tqdm
